Adicionando refatoração para Produto e adição de testes unitários

+++ src/main/java/com/example/demo/service/ProdutoService.java
@@ -6,6 +6,7 @@ import com.example.demo.enums.DemoApiCodeEnum;
 import com.example.demo.exception.ApiException;
 import com.example.demo.repositories.ProdutoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
@@ -18,48 +19,50 @@ public class ProdutoService {
     @Autowired
     private ProdutoRepository produtoRepository;
 
+    private static final Integer DEFAULT_QUANTITY = 1;
+
     public Produto createProduct (ProdutoDTO produtoDTO) {
-        if (produtoRepository.getProductByName(produtoDTO.getNome()).isPresent()) {
+        if (produtoRepository.findByNome(produtoDTO.getNome()).isPresent()) {
             throw new ApiException(DemoApiCodeEnum.API_ERROR_EXISTENT_PRODUCT, "Produto já cadastrado. Tente novamente");
         }
 
+        Produto produto = toEntity(produtoDTO);
+        return produtoRepository.save(produto);
+    }
 
-        Produto produto = Produto.builder()
+    private static Produto toEntity(ProdutoDTO produtoDTO) {
+        return Produto.builder()
                 .nome(produtoDTO.getNome())
                 .valor(produtoDTO.getValor())
-                .quantidade(50)
+                .quantidade(DEFAULT_QUANTITY)
                 .build();
-
-        produtoRepository.save(produto);
-        return produto;
     }
 
     public void deleteProduct(Long id) {
-        try {
-            produtoRepository.deleteById(id);
-        } catch (Exception e) {
-            throw new ApiException(DemoApiCodeEnum.API_ERROR_COULDNT_DELETE_PRODUCT, "Não foi possível deletar o produto.");
+        if (!produtoRepository.existsById(id)) {
+            throw new ApiException(DemoApiCodeEnum.API_ERROR_PRODUCT_DOESNT_EXIST, "Produto não existente na base de dados");
         }
+        produtoRepository.deleteById(id);
     }
 
     public Produto getProductById(Long id) {
-        return produtoRepository.getProductById(id).orElseThrow(() -> new ApiException(DemoApiCodeEnum.API_ERROR_PRODUCT_DOESNT_EXIST, "Produto não existe"));
+        return produtoRepository.findById(id).orElseThrow(() -> new ApiException(DemoApiCodeEnum.API_ERROR_PRODUCT_DOESNT_EXIST, "Produto não existe"));
     }
 
-    public List<Produto> getAllProducts() {
-        return produtoRepository.getAllProducts();
+    public List<Produto> getAllProducts(Pageable pageable) {
+        return produtoRepository.findAll();
     }
 
     public boolean validateProduct(String name) {
-        return produtoRepository.getProductByName(name).isPresent();
+        return produtoRepository.findByNome(name).isPresent();
     }
 
     public BigDecimal getProductPriceByName(String name) {
-        return produtoRepository.getProductByName(name).orElseThrow(() -> new ApiException(DemoApiCodeEnum.API_ERROR_PRODUCT_DOESNT_EXIST, "Produto não existe")).getValor();
+        return produtoRepository.findByNome(name).orElseThrow(() -> new ApiException(DemoApiCodeEnum.API_ERROR_PRODUCT_DOESNT_EXIST, "Produto não existe")).getValor();
     }
 
     public Produto getProductByName(String name) {
-        return produtoRepository.getProductByName(name).orElseThrow(() -> new ApiException(DemoApiCodeEnum.API_ERROR_PRODUCT_DOESNT_EXIST, "Produto não existe"));
+        return produtoRepository.findByNome(name).orElseThrow(() -> new ApiException(DemoApiCodeEnum.API_ERROR_PRODUCT_DOESNT_EXIST, "Produto não existe"));
 
     }
